Refactor: Extract outline parsing to core module
Move outline parsing logic into a new core module to improve code organization and reusability. This change centralizes the parsing functionality, making it accessible from both the main application and the worker thread. The `useOutline` hook and the outline worker now delegate parsing to this new module.
